job-listings-api/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
pytest==6.2.4
requests==2.26.0
python-dotenv==0.19.1